+++ test/pipeline/test_FeatureUnion.py
@@ -10,9 +10,9 @@ from dtoolkit.pipeline import make_pipeline
 from dtoolkit.pipeline import make_union
 from dtoolkit.transformer import GetTF
 from dtoolkit.transformer import OneHotEncoder
-from test.transformer.conftest import df_iris
-from test.transformer.conftest import df_label
-from test.transformer.conftest import df_mixed
+from test.transformer.data import df_iris
+from test.transformer.data import df_label
+from test.transformer.data import df_mixed
 
 
 # include `make_union`
