+++ test/pipeline/test_Pipeline.py
@@ -15,10 +15,10 @@ from dtoolkit.transformer import FilterInTF
 from dtoolkit.transformer import GetTF
 from dtoolkit.transformer import QueryTF
 from dtoolkit.transformer import RavelTF
-from test.transformer.conftest import df_iris
-from test.transformer.conftest import df_mixed
-from test.transformer.conftest import feature_names
-from test.transformer.conftest import s
+from test.transformer.data import df_iris
+from test.transformer.data import df_mixed
+from test.transformer.data import feature_names
+from test.transformer.data import s
 
 
 # include `make_pipeline`
