+++ test/transformer/pandas/test_AssignTF.py
@@ -1,7 +1,7 @@
 from functools import partial
 
 from dtoolkit.transformer import AssignTF
-from test.transformer.conftest import df_period
+from test.transformer.data import df_period
 
 
 def period(df, regex):
