+++ test/transformer/pandas/test_EvalTF.py
@@ -1,8 +1,8 @@
 from pandas.testing import assert_series_equal
 
 from dtoolkit.transformer import EvalTF
-from test.transformer.conftest import df_iris
-from test.transformer.conftest import feature_names
+from test.transformer.data import df_iris
+from test.transformer.data import feature_names
 
 
 def test_work():
