+++ test/transformer/pandas/test_QueryTF.py
@@ -1,8 +1,8 @@
 from pandas.testing import assert_frame_equal
 
 from dtoolkit.transformer import QueryTF
-from test.transformer.conftest import df_iris
-from test.transformer.conftest import feature_names
+from test.transformer.data import df_iris
+from test.transformer.data import feature_names
 
 
 def test_greater_symbol():
