+++ test/transformer/pandas/test_SelectDtypesTF.py
@@ -2,9 +2,9 @@ import pytest
 from pandas.testing import assert_frame_equal
 
 from dtoolkit.transformer import SelectDtypesTF
-from test.transformer.conftest import df_iris
-from test.transformer.conftest import df_label
-from test.transformer.conftest import df_mixed
+from test.transformer.data import df_iris
+from test.transformer.data import df_label
+from test.transformer.data import df_mixed
 
 
 @pytest.mark.parametrize(
